requirements.txt
Django==5.0
django-rest-framework==0.1.0
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0